+++ AnkiDroid/src/main/java/com/ichi2/anki/servicelayer/ScopedStorageService.kt
@@ -20,8 +20,8 @@ package com.ichi2.anki.servicelayer
 import android.content.Context
 import android.os.Build
 import android.os.Environment
-import com.ichi2.anki.CollectionHelper
 import com.ichi2.anki.common.preferences.sharedPrefs
+import com.ichi2.anki.common.storage.CollectionHelper
 import com.ichi2.anki.ui.windows.managespace.isInsideDirectoriesRemovedWithTheApp
 import com.ichi2.utils.Permissions
 import timber.log.Timber
