+++ AnkiDroid/src/test/java/com/ichi2/anki/StorageDecisionGateTest.kt
@@ -5,8 +5,9 @@ package com.ichi2.anki
 import androidx.core.content.edit
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.common.preferences.sharedPrefs
+import com.ichi2.anki.common.storage.CollectionHelper
+import com.ichi2.anki.common.storage.StorageDecision
 import com.ichi2.anki.exception.StorageNotConfiguredException
-import com.ichi2.anki.storage.StorageDecision
 import org.junit.After
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -16,8 +17,8 @@ import kotlin.test.assertFailsWith
 
 /**
  * Proves the storage-decision gate in [CollectionManager.ensureOpenInner] is wired up.
- * [CollectionHelper.storageDecision] is [StorageDecision.Decided] once the user has chosen
- * where the collection is stored ([CollectionHelper.PREF_COLLECTION_PATH] is set); tests force
+ * [com.ichi2.anki.common.storage.CollectionHelper.storageDecision] is [StorageDecision.Decided] once the user has chosen
+ * where the collection is stored ([com.ichi2.anki.common.storage.CollectionHelper.PREF_COLLECTION_PATH] is set); tests force
  * [StorageDecision.Undecided] via the test override.
  */
 @RunWith(AndroidJUnit4::class)
