+++ AnkiDroid/src/test/java/com/ichi2/anki/preferences/PreferencesScreenshotTest.kt
@@ -4,9 +4,9 @@ package com.ichi2.anki.preferences
 
 import androidx.preference.Preference
 import androidx.test.core.app.ActivityScenario
-import com.ichi2.anki.CollectionHelper
 import com.ichi2.anki.R
 import com.ichi2.anki.ScreenshotTest
+import com.ichi2.anki.common.storage.CollectionHelper
 import com.ichi2.anki.settings.Prefs
 import org.junit.Test
