+++ AnkiDroid/src/test/java/com/ichi2/anki/startup/SetupStorageTest.kt
@@ -4,9 +4,9 @@ package com.ichi2.anki.startup
 
 import androidx.core.content.edit
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.ichi2.anki.CollectionHelper
 import com.ichi2.anki.RobolectricTest
 import com.ichi2.anki.common.preferences.sharedPrefs
+import com.ichi2.anki.common.storage.CollectionHelper
 import com.ichi2.anki.exception.StorageNotConfiguredException
 import org.junit.After
 import org.junit.Test
