+++ anki-common/src/main/java/com/ichi2/anki/common/storage/AnkiDroidFolder.kt
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package com.ichi2.anki.storage
+package com.ichi2.anki.common.storage
 
 /**
  * Where AnkiDroid stores its collection.
